requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.0.0
python-dateutil>=2.8.0
streamlit>=1.29.0
plotly>=5.18.0
streamlit-autorefresh>=1.0.1
smartapi-python>=1.3.0
pyotp>=2.8.0
websocket-client>=1.5.0
